### 2024-06-20 #6910 @editorialbot
[PRE REVIEW]: rOPTRAM: Deriving Soil Moisture from Satellite Imagery in R
**Submitting author:** @micha-silver (<a href="http://orcid.org/0000-0002-1128-1325">Micha Silver</a>)
**Repository:** https://github.com/ropensci/rOPTRAM
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
