### 2024-06-20 #6910 @kthyng
@micha-silver Can you try to fix the error in compiling the paper? On a quick glance it doesn't look like your paper is following the example paper syntax correctly: https://joss.readthedocs.io/en/latest/example_paper.html#example-paper
