### 2024-06-20 #6910 @kthyng
Also please remove the backticks around the package name and "R" for it to compile correctly — we can try this first to see if it fixes the issue.
