### 2024-06-20 #6910 @micha-silver
@kthyng
In the example paper, R package names are surrounded by backticks. i.e.:

```
# Statement of need
`Gala` is an Astropy-affiliated Python package for galactic dynamics.
```
??
