### 2024-06-20 cafe47a
Changes structure of affiliations in paper.md
