+++ inst/paper/paper.md
@@ -1,19 +1,22 @@
 ---
-title: "`rOPTRAM`: Implementation of the OPTRAM Algorithm in `R`"
-author:
+title: "rOPTRAM: Implementation of the OPTRAM Algorithm in R"
+authors:
 - name: Micha Silver
   orcid: "0000-0002-1128-1325"
-  affiliation: Ben Gurion University, Beer Sheva, Israel
+  affiliation: 1
   corresponding: yes
 - name: Arnon Karnieli
   orcid: "0000-0001-8065-9793"
-  affiliation: Ben Gurion University, Beer Sheva, Israel
+  affiliation: 1
+affiliations:
+ - name: Blaustein Institutes for Desert Research, Ben Gurion University, Israel
+   index: 1
 date: "07 June 2024"
 output:
-  word_document: default
+  pdf_document: default
   html_document:
     df_print: paged
-  pdf_document: default
+  word_document: default
 bibliography: bibliography.bib
 tags:
 - soil moisture
@@ -26,7 +29,7 @@ tags:
 
 While drylands supply a livelihood to much of the world's rural population, these arid and semi-arid areas are under increased pressure due to both growing demand and the current climate crisis. Maintaining a sustainable food source for rural populations depends on reliable grazing, and the quality of grazing is in turn determined by soil moisture. Thus soil moisture must be monitored to ensure a sustainable food supply. Classic soil moisture monitoring methods rely on sensors, inserted into the soil, that give accurate measurements at high temporal resolution, but at a point location.
 
-However, addressing the needs of populations that depend on extensive grazing lands requires a regional scale soil moisture assessment. Point measurements, albeit accurate, do not afford the needed information to prepare for, or mitigate drought events at regional scale. To this end, remote sensing methods to estimate soil moisture have been developed. Among them, the OPTRAM model has been shown to accurately determine soil moisture over large areas. The `rOPTRAM` package in `R` has implemented that model allowing researchers and practitioners to monitor grazing potential at regional scale and long time intervals.
+However, addressing the needs of populations that depend on extensive grazing lands requires a regional scale soil moisture assessment. Point measurements, albeit accurate, do not afford the needed information to prepare for, or mitigate drought events at regional scale. To this end, remote sensing methods to estimate soil moisture have been developed. Among them, the OPTRAM model has been shown to accurately determine soil moisture over large areas. The rOPTRAM package in R has implemented that model allowing researchers and practitioners to monitor grazing potential at regional scale and long time intervals.
 
 # Introduction
 
@@ -39,11 +42,11 @@ The overarching goal of the current project was to program OPTRAM using the open
 
 # Algorithm
 
-`rOPTRAM` produces a large dataset of pixel values of two satellite-based raster layers: a vegetation index (VI), such as Normalized Difference Vegetation Index (NDVI) and the STR layer. All pairs of pixel values, at all acquired image dates are plotted as a scatter plot. Then regression lines are extracted at both the upper ("wet") and lower ("dry") bounds of the scatter plot. The slopes and intercepts of these two regression lines are the model coefficients, used to derive a spatially explicit soil moisture map. This soil moisture map is calculated, following @sadeghi_optical_2017, using  \autoref{eq:sm}. In his original work, @sadeghi_optical_2017 used a visual examination of the scatterplot to locate the trapezoid edges.
[…]
