### 2024-06-20 4754fc0
Fixes affiliations in paper.md
