### 2024-06-20 c49e585
Feixes to yaml header in paper.md
