+++ inst/paper/paper.md
@@ -1,36 +1,36 @@
 ---
-title: "`rOPTRAM`: Implementation of the OPTRAM Algorithm in `R`"
+title: 'rOPTRAM: Implementation of the OPTRAM Algorithm in R'
+tags:
+  - soil moisture
+  - remote sensing
+  - Sentinel-2
+  - R
 author:
-- name: Micha Silver
-  orcid: "0000-0002-1128-1325"
-  affiliation: 1
-  corresponding: yes
-- name: Dong Zhe
-  orcid: 0009-0007-5298-8355
-  affiliation: 1
-- name: Arnon Karnieli
-  orcid: "0000-0001-8065-9793"
-  affiliation: 1
+  - name: Micha Silver
+    orcid: 0000-0002-1128-1325
+    affiliation: 1
+    corresponding: true
+  - name: Dong Zhe
+    orcid: 0009-0007-5298-8355
+    affiliation: 1
+  - name: Arnon Karnieli
+    orcid: "0000-0001-8065-9793"
+    affiliation: 1
 affiliations:
-  name: Blaustein Institutes for Desert Research, Ben Gurion University, Israel
-  index: 1
+  - name: Blaustein Institutes for Desert Research, Ben Gurion University, Israel
+    index: 1
 date: "07 June 2024"
 output:
   html_document:
-    df_print: paged
+  pdf_print: paged
   pdf_document: default
   word_document: default
 bibliography: bibliography.bib
-tags:
-- soil moisture
-- remote sensing
-- "Sentinel-2"
-- R
 ---
 
 # Summary
 
-Soil moisture at the Earth’s surface is essential in many hydrological, pedological, and biological processes. Earth observations from satellites have been recognized as the most efficient and reliable means for assessing soil water content.  The satellite-derived OPTRAM model has been shown to determine soil moisture over large areas accurately. The overarching goal of the current project was to program the original OPTRAM algorithm, with some improvements, using the open-source R language. The spectral data were obtained from different spaceborne systems to calculate several vegetation indices, and the model was adjusted to various SM conditions. The `rOPTRAM` package allows researchers and practitioners to monitor soil moisture contents at a regional scale and long-time intervals.
+Soil moisture at the Earth’s surface is essential in many hydrological, pedological, and biological processes. Earth observations from satellites have been recognized as the most efficient and reliable means for assessing soil water content.  The satellite-derived OPTRAM model has been shown to determine soil moisture over large areas accurately. The overarching goal of the current project was to program the original OPTRAM algorithm, with some improvements, using the open-source ``R`` language. The spectral data were obtained from different spaceborne systems to calculate several vegetation indices, and the model was adjusted to various SM conditions. The ``rOPTRAM`` package allows researchers and practitioners to monitor soil moisture contents at a regional scale and long-time intervals.
 
 # Statement of Need
 
@@ -38,16 +38,16 @@ Soil moisture (SM) at the Earth’s surface is essential in many hydrological, p
 
 Less than a decade ago, @sadeghi_optical_2017 showed that shortwave infrared (SWIR) Transformed Reflectance (STR) is applicative to assess SM. They developed a novel physical-based model, OPtical TRApezoid Model (OPTRAM) using designated spectral bands of remote sensing imagery. OPTRAM was recently validated [@longo-minnolo_stand-alone_2022] to address the need to estimate SM over vast areas in watershed and regional scales. Consequently the model can now be applied to various earth observation systems, such as Sentinel-2 or Landsat, with visible, near-infrared, and SWIR bands. [@ambrosone_retrieving_2020; @dubinin_using_2020].
 
[…]
