### 2024-06-20 312bc5c
Edits to paper.md for successful compilation in docker
