### 2024-06-20 #6910 @micha-silver
I made a few changes to the yaml header and then successfully compiled (using the suggested docker image). The only problem I see now is that the two included images do not appear in the final pdf  :-(
