### 2024-06-20 c1d8b20
Fix image links and labels
