### 2024-06-20 #6910 @micha-silver
@kthyng I found the correct format for figures and fixed paper.md accordingly. Now the pdf compiles successfully including images.
Thanks for your tips.
