### 2024-06-21 #6910 @danielskatz
> @danielskatz the author changed them in the paper, do they matter in the issue name?

I don't think so, but you can manually edit the issue name if you want
