### 2024-06-21 f8a7469
Adds doi to CDSE citation
