+++ inst/paper/bibliography.bib
@@ -1,5 +1,5 @@
 @article{maestre_grazing_2022,
-	author = {Fernando T. Maestre and Yoann Le Bagousse-Pinguet and Manuel Delgado-Baquerizo and David J. Eldridge and Hugo Sáiz and Miguel Berdugo and Beatriz Gozalo and Victoria Ochoa and Emilio Guirado and Miguel García-Gómez and Enrique Valencia and Xiaobing Zhou and Brajesh K. Singh and Nicolas Gross and Betty Mendoza and César Plaza de Carlos and Paloma Díaz-Martínez and Ana Rey and Hang-Wei Hu and Ji-Zheng He and Jun-Tao Wang and Chongfeng Bu and Anika Lehmann and Matthias C. Rillig and Simone Cesarz and Nico Eisenhauer and Jaime Martínez-Valderrama and Eduardo Moreno-Jiménez and O. Salas and Mehdi Abedi and Negar Ahmadian  and Concepción L. Alados and Rafaella Canessa and Valeria Aramayo and Fateh Amghar and Tulio Arredondo and Rodrigo J. Ahumada and Khadijeh Bahalkeh and Farah Ben Salem and Niels Blaum and Bazartseren Boldgiv and Matthew A. Browker and Donaldo Bran and Andrea P. Castillo-Monroy and Helena Castro and Ignacio Castro and Patricio Castro-Quezada and Roukaya Chibani and Abel A. Conceição and Courtney M. Currier and Anthony Darrouzet-Nardi and Florian Jeltsch and Balázs Deák and David Donoso and Andrew Dougill and Jorge Durán and Batdegleg Erdenetsetseg and Carlos Iván Espinosa and Alex Fajardo and Mohammad Farzam and Daniela Ferrante and Anke S. K. Frank and Anke Jentsch and L. H. Fraser and Laureano Gherardi and Aaron Greenville and Carlos A. Guerra and Elizabeth Gusmán Montalván and Rosa M. Hernández Hernández and Norbert Hölzel and E. Huber-Sannwald and Frederic M. Hughes and Oswaldo Jadán and Kudzai Farai Kaseke and Melanie Köbel and Jesica E. Koopman and Cintia Leder and Anja Linstädter and Peter C. Le Roux and Xinkai Li and Pierre Liancourt and Alexandra Rodríguez-Pereiras and Jushan Liu and Michelle A. Low and Gillian Maggs Kölling and Thulani P. Makhalanyane and Oumarou Malam Issa and Antonio J. Manzaneda and Eugene Marais and Juan P. Mora and Gerardo Moreno and Seth M. Munson and Víctor Rolo and Alice Nunes and Gabriel Oliva and Gastón Oñatibia and Marco O. D. Pivari and Yolanda Pueyo and R Emiliano Quiroga and Soroor Rahmanian and Sasha C. Reed and P.J. Rey and Benoit Richard and Juan G. Rubalcaba and Jan C. Ruppert and Ayman Salah and Max A. Schuchardt and Sedona Spann and Ilan Stavi and Colton R. A. Stephens and Anthony Swemmer and Juan J. Gaitán and Alberto L. Teixido and Andrew Thomas and Heather L. Throop and Katja Tielbörger and Samantha Travers and James Val and Orsolya Valkó and Liesbeth van den Brink and Sergio Velasco Ayuso and Frederike Velbert and Sergio Asencio and Wanyoike Wamiti and Deli Wang and Lixin Wang and Glenda M. Wardle and Laura Yahdjian and Eli Zaady and Zhang Yuanming},
[…]
