### 2024-06-21 fc89fc5
Removes r-cran bibliography entry
