+++ inst/paper/bibliography.bib
@@ -99,16 +99,6 @@
 	pages = {9143},
 }
 
-
-@Manual{r_core_2022,
-	title = {R: A Language and Environment for Statistical Computing},
-	author = {{R Core Team}},
-	organization = {R Foundation for Statistical Computing},
-	address = {Vienna, Austria},
-	year = {2022},
-	url = {https://www.R-project.org/},
-}
-
 @article{zhang_estimation_2016,
 	title = {Estimation of {Soil} {Moisture} from {Optical} and {Thermal} {Remote} {Sensing}: {A} {Review}},
 	volume = {16},
