### 2024-06-21 #6910 @micha-silver
Hi: @kthyng @danielskatz
I found the missing doi for the `CDSE` , and removed the citation to R-cran (No doi available since it's a project, not a single document).
I'll choose some additional potential reviewers and post here soon.
Thanks, Micha
