### 2024-06-25 f238c83
Defines env variables (SCIHUB_CLIENTID and SCIHUB_SECRET) for github actions
