+++ .github/workflows/R-CMD-check.yaml
@@ -1,10 +1,10 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
-on:
-  push:
-    branches: [main, master]
-  pull_request:
-    branches: [main, master]
+#on:
+#  push:
+#    branches: [main, master]
+#  pull_request:
+#    branches: [main, master]
 
 name: R-CMD-check
 
@@ -29,7 +29,9 @@ jobs:
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
-
+      clientid: ${{ secrets.SCIHUB_CLIENTID }}
+      secret: ${{ secrets.SCIHUB_SECRET }}
+ 
     steps:
       - uses: actions/checkout@v4
 
