+++ .github/workflows/R-CMD-check.yaml
@@ -1,6 +1,6 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
-#on:
+on:
 #  push:
 #    branches: [main, master]
 #  pull_request:
@@ -31,7 +31,7 @@ jobs:
       R_KEEP_PKG_SOURCE: yes
       clientid: ${{ secrets.SCIHUB_CLIENTID }}
       secret: ${{ secrets.SCIHUB_SECRET }}
- 
+
     steps:
       - uses: actions/checkout@v4
 
