### 2024-06-25 63a136e
Corrects 'on: push:' in R-CMD_check.yaml
