+++ .github/workflows/R-CMD-check.yaml
@@ -1,7 +1,7 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
-#  push:
+  push:
 #    branches: [main, master]
 #  pull_request:
 #    branches: [main, master]
