### 2024-06-25 2b8685a
Corrects R-CMD-check.yaml, adds environment: CDSE_environment
