### 2024-06-25 06222bf
Changes R-CMD-check.yaml to run at workflow-dispatch
