+++ .github/workflows/R-CMD-check.yaml
@@ -1,8 +1,8 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
-on:
-  push:
-    branches: [main, dev-options]
+on: workflow_dispatch
+#  push:
+#    branches: [main, dev-options]
 #  pull_request:
 #    branches: [main, master]
 
@@ -25,8 +25,6 @@ jobs:
           - {os: ubuntu-latest,   r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest,   r: 'release'}
           - {os: ubuntu-latest,   r: 'oldrel-1'}
-
-    environment: CDSE_environment
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
