### 2024-06-25 448f6fe
Adds save_creds=FALSE to test-acquire_s2 test
