### 2024-06-25 2a2559c
Adds env definition to steps in R-CMD-check.yaml
