+++ .github/workflows/R-CMD-check.yaml
@@ -49,11 +49,17 @@ jobs:
           use-public-rspm: true
 
       - uses: r-lib/actions/setup-r-dependencies@v2
+        env:
+          clientid: ${{ secrets.SCIHUB_CLIENTID }}
+          secret: ${{ secrets.SCIHUB_SECRET }}
         with:
           extra-packages: any::rcmdcheck
           needs: check
 
       - uses: r-lib/actions/check-r-package@v2
+        env:
+          clientid: ${{ secrets.SCIHUB_CLIENTID }}
+          secret: ${{ secrets.SCIHUB_SECRET }}
         with:
           upload-snapshots: true
           build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
