### 2024-06-25 f419bbd
Replaces deprecated GetArchiveImage in acquire_utilities.R
