### 2024-06-25 #6910 @kthyng
I updated the title of the issue so it matches the JOSS paper title exactly.
