### 2024-06-25 #6910 @kthyng
and yeah, the license tool is imperfect, but helpful.
