### 2024-06-25 9ddd160
Adds github action yaml for one platform
