+++ .github/workflows/ubuntu-latest-check.yaml
@@ -6,7 +6,7 @@ on: workflow_dispatch
 #  pull_request:
 #    branches: [main, master]
 
-name: R-CMD-check
+name: One-platform
 
 permissions:
   actions: write
@@ -16,20 +16,13 @@ permissions:
   statuses: read
 
 jobs:
-  R-CMD-check:
-    runs-on: ${{ matrix.config.os }}
+  One-platform:
+    runs-on: ubuntu-latest
 
-    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
+    name: One platform
 
     strategy:
       fail-fast: false
-      matrix:
-        config:
-          - {os: macos-latest,   r: 'release'}
-          - {os: windows-latest, r: 'release'}
-          - {os: ubuntu-latest,   r: 'devel', http-user-agent: 'release'}
-          - {os: ubuntu-latest,   r: 'release'}
-          - {os: ubuntu-latest,   r: 'oldrel-1'}
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
@@ -44,7 +37,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r@v2
         with:
-          r-version: ${{ matrix.config.r }}
+          r-version: 'release'
           clientid: ${{ secrets.SCIHUB_CLIENTID }}
           secret: ${{ secrets.SCIHUB_SECRET }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
