### 2024-06-25 2fd7b27
Removes duplicate ubuntu-latest-check
