+++ .github/workflows/ubuntu-latest-check.yml
@@ -1,57 +1,0 @@
-
-on: workflow_dispatch
-#  push:
-#    branches: [main, dev-options]
-#  pull_request:
-#    branches: [main, master]
-
-name: One-platform
-
-permissions:
-  actions: write
-  checks: write
-  issues: write
-  contents: read
-  statuses: read
-
-jobs:
-  One-platform:
-    runs-on: ubuntu-latest
-
-    name: One platform
-
-    strategy:
-      fail-fast: false
-
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      R_KEEP_PKG_SOURCE: yes
-      clientid: ${{ secrets.SCIHUB_CLIENTID }}
-      secret: ${{ secrets.SCIHUB_SECRET }}
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: r-lib/actions/setup-pandoc@v2
-
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: 'release'
-          clientid: ${{ secrets.SCIHUB_CLIENTID }}
-          secret: ${{ secrets.SCIHUB_SECRET }}
-          http-user-agent: ${{ matrix.config.http-user-agent }}
-          use-public-rspm: true
-
-      - uses: r-lib/actions/setup-r-dependencies@v2
-        with:
-          clientid: ${{ secrets.SCIHUB_CLIENTID }}
-          secret: ${{ secrets.SCIHUB_SECRET }}
-          extra-packages: any::rcmdcheck
-          needs: check
-
-      - uses: r-lib/actions/check-r-package@v2
-        with:
-          clientid: ${{ secrets.SCIHUB_CLIENTID }}
-          secret: ${{ secrets.SCIHUB_SECRET }}
-          upload-snapshots: true
-          build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
