### 2024-06-25 37f81aa
Edits to ubuntu-latest-check.yaml
