+++ .github/workflows/ubuntu-latest-check.yaml
@@ -5,7 +5,7 @@ on: workflow_dispatch
 #  pull_request:
 #    branches: [main, master]
 
-name: One-platform
+name: One-platform-check
 
 permissions:
   actions: write
@@ -15,10 +15,10 @@ permissions:
   statuses: read
 
 jobs:
-  One-platform:
+  One-platform-check:
     runs-on: ubuntu-latest
 
-    name: One platform
+    name: One platform check
 
     strategy:
       fail-fast: false
@@ -37,8 +37,6 @@ jobs:
       - uses: r-lib/actions/setup-r@v2
         with:
           r-version: 'release'
-          clientid: ${{ secrets.SCIHUB_CLIENTID }}
-          secret: ${{ secrets.SCIHUB_SECRET }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
 
