### 2024-06-26 8f9c4e8
Edits ubuntu-latest-check.yaml
