+++ .github/workflows/ubuntu-latest-check.yaml
@@ -41,15 +41,17 @@ jobs:
           use-public-rspm: true
 
       - uses: r-lib/actions/setup-r-dependencies@v2
-        with:
+        secrets:
           clientid: ${{ secrets.SCIHUB_CLIENTID }}
           secret: ${{ secrets.SCIHUB_SECRET }}
+        with:
           extra-packages: any::rcmdcheck
           needs: check
 
       - uses: r-lib/actions/check-r-package@v2
-        with:
+        secrets:
           clientid: ${{ secrets.SCIHUB_CLIENTID }}
           secret: ${{ secrets.SCIHUB_SECRET }}
+        with:
           upload-snapshots: true
           build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
