### 2024-06-26 9cc11c3
Update ubuntu-latest-check.yaml
Edits ubuntu-latest-check.yaml
