### 2024-06-26 b9cf410
Update ubuntu-latest-check.yaml
Edits again ubuntu-latest-check.yaml
