### 2024-06-26 da0a35a
Update ubuntu-latest-check.yaml
