### 2024-06-27 67aa9ab
fix: use same name as env var in workflow file
