### 2024-06-27 8fa6b2d
fixes env variable names also  in test-acquire_s2.R
