### 2024-06-27 f5bc6d9
Fixes env variable names also in R-CMD-check.yaml
