+++ .github/workflows/R-CMD-check.yaml
@@ -34,8 +34,8 @@ jobs:
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
-      clientid: ${{ secrets.SCIHUB_CLIENTID }}
-      secret: ${{ secrets.SCIHUB_SECRET }}
+      SCIHUB_CLIENTID: ${{ secrets.SCIHUB_CLIENTID }}
+      SCIHUB_SECRET: ${{ secrets.SCIHUB_SECRET }}
 
     steps:
       - uses: actions/checkout@v4
@@ -45,21 +45,24 @@ jobs:
       - uses: r-lib/actions/setup-r@v2
         with:
           r-version: ${{ matrix.config.r }}
-          clientid: ${{ secrets.SCIHUB_CLIENTID }}
-          secret: ${{ secrets.SCIHUB_SECRET }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
+        env:
+          SCIHUB_CLIENTID: ${{ secrets.SCIHUB_CLIENTID }}
+          SCIHUB_SECRET: ${{ secrets.SCIHUB_SECRET }}
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          clientid: ${{ secrets.SCIHUB_CLIENTID }}
-          secret: ${{ secrets.SCIHUB_SECRET }}
           extra-packages: any::rcmdcheck
           needs: check
+        env:
+          SCIHUB_CLIENTID: ${{ secrets.SCIHUB_CLIENTID }}
+          SCIHUB_SECRET: ${{ secrets.SCIHUB_SECRET }}
 
       - uses: r-lib/actions/check-r-package@v2
         with:
-          clientid: ${{ secrets.SCIHUB_CLIENTID }}
-          secret: ${{ secrets.SCIHUB_SECRET }}
           upload-snapshots: true
           build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
+        env:
+          SCIHUB_CLIENTID: ${{ secrets.SCIHUB_CLIENTID }}
+          SCIHUB_SECRET: ${{ secrets.SCIHUB_SECRET }}
