### 2024-06-28 89ad29c
Removes test for CI veg index in test-utilities.R
