+++ tests/testthat/test-acquire_s2.R
@@ -11,8 +11,7 @@ test_that("optram_acquire_s2 returns list", {
   if (secret == "") secret <- NULL
   boa_list <- optram_acquire_s2(aoi, from_date, to_date,
                                 clientid = clientid, secret = secret,
-                                save_creds = FALSE,
-                                remote = "scihub", SWIR_band = 11)
+                                save_creds = FALSE)
   expect_true(inherits(boa_list, "list"))
 })
 
