### 2024-06-28 aea8bb3
Removes check_swir_band() function from utilities.R
