### 2024-06-28 f042f48
Minor updates to documentation of acquire_utilities.R
