+++ R/optram_acquire_s2.R
@@ -101,11 +101,9 @@
 #' to_date <- "2019-04-30"
 #' aoi <- sf::st_read(system.file("extdata",
 #'                               "lachish.gpkg", package = 'rOPTRAM'))
+#' aoi <- sf::st_as_sf(sf::st_union(aoi))
 #' s2_file_list <- optram_acquire_s2(aoi,
-#'                                  from_date, to_date,
-#'                                  remote = "scihub",
-#'                                  veg_index = "SAVI",
-#'                                  SWIR_band = 11)
+#'                                  from_date, to_date)
 #' }
 
 optram_acquire_s2 <- function(
@@ -119,9 +117,6 @@ optram_acquire_s2 <- function(
   # Avoid "no visible binding for global variable" NOTE
   scihub <- openeo <- NULL
 
-  # Get the SWIR_band and remote from package options
-  remote <- getOption("optram.remote")
-  SWIR_band <- getOption("optram.SWIR_band")
   # Pre flight checks...
   if (!check_aoi(aoi)) return(NULL)
   # Ensure that aoi is single POLYGON or MULTIPOLYGON feature (and sf NOT sfc)
