+++ tests/testthat/test-acquire_s2.R
@@ -1,8 +1,8 @@
 skip_if_not_installed("CDSE")
 
 test_that("optram_acquire_s2 returns list", {
-  from_date <- "2023-03-01"
-  to_date <- "2023-04-30"
+  from_date <- "2022-03-01"
+  to_date <- "2022-04-30"
   aoi <- sf::st_read(system.file("extdata",
                                  "lachish.gpkg", package = "rOPTRAM"))
   clientid <- Sys.getenv("SCIHUB_CLIENTID")
