### 2024-06-28 61997c2
Fixes loop over dates in acquire_scihub()
