### 2024-07-01 8819780
Bump to v0.3.0 with options and corrected vignettes
