+++ R/zzz.R
@@ -20,7 +20,7 @@
     optram.max_tbl_size = 1e+6,  # depends on computer resources
     optram.rm.low.vi    = FALSE,
     optram.rm.hi.str    = FALSE,
-    optram.plot_colors = "no",  # "colors"
+    optram.plot_colors = "no",  # "density", "contours", "feature"
     optram.feature_col = "ID",  # Which column contains numeric ID's of features
     optram.edge_points = TRUE   # whether to add edge points to the plot
   )
