+++ vignettes/rOPTRAM_acquiring.Rmd
@@ -37,7 +37,7 @@ To utilize the CDSE package for accessing Copernicus data, follow these steps:
 Ensure you have the `CDSE` and `jsonlite` packages installed:
 
 ```{r setup, results='hide', eval=FALSE}
-remotes::install_github("ropensci/rOPTRAM")
+#remotes::install_github("ropensci/rOPTRAM")
 library(rOPTRAM)
 if (!require("CDSE")) install.packages("CDSE", dependencies = TRUE)
 if (!require("jsonlite")) install.packages("jsonlite", dependencies = TRUE)
