+++ man/optram_options.Rd
@@ -41,7 +41,7 @@ rm.low.vi        \tab FALSE \tab TRUE \cr
 rm.hi.str        \tab FALSE \tab TRUE \cr
 plot_colors      \tab "no"  \tab "no" = uniform green color for all points \cr
 \tab  \tab "features" = points colored by aoi features \cr
-\tab  \tab "colors" = points colored by point density \cr
+\tab  \tab "density" = points colored by point density \cr
 \tab  \tab "contours" = plots density contour lines \cr
 feature_col      \tab "ID" \tab string, name of \emph{numeric} column \cr
 \tab  \tab that contains feature ID's for coloring plot \cr
