+++ vignettes/rOPTRAM_trapezoid_methods.Rmd
@@ -41,7 +41,7 @@ The fitting method is chosen by setting the `trapezoid_method` option using the 
 
 
 ``` r
-#remotes::install_github("ropensci/rOPTRAM")
+remotes::install_github("ropensci/rOPTRAM")
 library("rOPTRAM")
 library("CDSE")
 library("jsonlite")
