### 2024-07-02 ef3bb5e
Adds new options to optram_options() function
