### 2024-07-02 #6910 @kthyng
@micha-silver A couple more comments:
* I suggest you change the title to be descriptive of what this algorithm does so someone who doesn't automatically know has some hint
* please try to cut down on the paper length. One suggestion is to move some of the extra length to the docs, if the content isn't already there. The paper is 50% too long.
