### 2024-07-03 3535898
Shortens the paper.md for JOSS
