+++ inst/paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'rOPTRAM: Implementation of the OPTRAM Algorithm in R'
+title: 'rOPTRAM: Deriving Soil Moisture from Satellite Imagery in R'
 tags:
   - soil moisture
   - remote sensing
@@ -25,13 +25,13 @@ bibliography: bibliography.bib
 
 # Summary
 
-Soil moisture at the Earth’s surface is essential in many hydrological, pedological, and biological processes. Earth observations from satellites have been recognized as the most efficient and reliable means for assessing soil water content.  The satellite-derived OPTRAM model has been shown to determine soil moisture over large areas accurately. The overarching goal of the current project was to program the original OPTRAM algorithm, with some improvements, using the open-source ``R`` language. The spectral data were obtained from different spaceborne systems to calculate several vegetation indices, and the model was adjusted to various SM conditions. The ``rOPTRAM`` package allows researchers and practitioners to monitor soil moisture contents at a regional scale and long-time intervals.
+Soil moisture at the Earth’s surface is essential in many hydrological, pedological, and biological processes. Earth observations from satellites have been recognized as the most efficient and reliable means for assessing soil water content.  The satellite-derived OPTRAM model has been shown to determine soil moisture over large areas accurately. Spectral data were obtained from different spaceborne systems to calculate several vegetation indices, and the model was adjusted to various SM conditions. The ``rOPTRAM`` package allows researchers and practitioners to monitor soil moisture contents at a regional scale and long-time intervals.
 
 # Statement of Need
 
-Soil moisture (SM) at the Earth’s surface is essential in many hydrological, pedological, and biological processes. Pasture is the most common land use, where grazing livestock spreads over about 77% of the global agricultural land, and support billions of inhabitants [@ramankutty_farming_2008; @maestre_grazing_2022]. Therefore SM is a critical resource in these regions for the growth of palatable plants, pasture quality, and rangeland health [@mbow_land_2021; @kusserow_desertification_2017]. A time series of SM measurements can serve to determine the quality and potential for grazing. By following seasonal variations in SM, forecasts for preferred grazing lands can be prepared. *In-situ* volumetric soil content is typically measured using Time Domain Reflectometer (TDR) sensors with a high temporal resolution [@kirkham_time_2014]. TDR sensors have been recognized to be very accurate, and can function for long periods. However these data are point measurements, and cannot cover the extensive areas needed for determining regional scale grazing potential. Cosmic Ray Neutron Scanner (CRNS) technology can also acquire SM measurements at high temporal resolution, with a larger spatial scale than TDR, covering up to 30 hectares. These latter instruments are, however, quite expensive, and less accurate than TDR spot measurements [@davies_optimal_2022; @schron_improving_2017]. To overcome the locality issue, Earth Observations from satellites have been accepted as the most efficient and reliable means for assessing the Earth’s resources, including SM, in isolated and inaccessible regions due to its large-scale coverage and repeatability for the last decades. 
[…]
