### 2024-07-03 #6910 @micha-silver
@kthyng
I changed the title to:  'rOPTRAM: Deriving Soil Moisture from Satellite Imagery in R'

and I shaved off some text. Now wordcount is 1400.  Still a bit too much. I'm not sure what you mean "moving extra to docs"?  I put into the paper some example code. This is more or less a repeat of what is in the function documentation. Should these examples be removed from the paper?
