### 2024-07-04 dfcf838
Adds "Month" to VI_STR_df for plotting monthly trajectory in scatter plot
