### 2024-07-04 f0baee8
Recreates BOA NDVI and STR in extdata folders to include tileID in downloaded files
