+++ man/calculate_vi.Rd
@@ -46,6 +46,6 @@ set scale_factor to 255.
 }
 \examples{
 img_stk <- terra::rast(system.file("extdata", "BOA",
-         "BOA_2023-01-20.tif", package = "rOPTRAM"))
+         "BOA_2023-01-20_T36RXV.tif", package = "rOPTRAM"))
 vi <- calculate_vi(img_stk)
 }
