### 2024-07-04 3890344
Fixers Feature_ID column in plot_vi_str_cloud() for plotting by features
