### 2024-07-04 0d94339
Also reverts VI_STR_sf$Month to integer (no longer a factor)
