### 2024-07-04 24030a6
Refactors Feature_ID creation in optram_ndvi_str()
