### 2024-07-04 106b400
Refactors Feature_ID creation in optram_ndvi_str()
