### 2024-07-04 5f33b29
Adds test for number of columns in VI_STR_df
