### 2024-07-04 5e5d53f
Edits to optram_ndvi_str() for getting Feature_ID
