### 2024-07-04 95c0fcb
Edits to optram_ndvi_str()
