### 2024-07-04 6df1efb
Refactors check for maximum table size
