### 2024-07-04 89eef02
Fixes idx problem in optram_ndvi_str()
