### 2024-07-05 #6910 @kthyng
@micha-silver was this already reviewed in rOpenSci?
