### 2024-07-05 #6910 @kthyng
@micha-silver I would move most or all of the algorithm and examples to your readme and/or docs. You can see what is expected in the paper here: https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain. Generally you should keep it high level. One example might be helpful if you think it will help a potential user understand what the software is used for.
