### 2024-07-06 #6910 @micha-silver
> @micha-silver was this already reviewed in rOpenSci?

@kthyng  Yes, the package passed review and is now hosted on ROpenSci:  [repo](https://github.com/ropensci/rOPTRAM)
