### 2024-07-06 #6910 @micha-silver
> @micha-silver I would move most or all of the algorithm and examples to your readme and/or docs. You can see what is expected in the paper here: https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain. Generally you should keep it high level. One example might be helpful if you think it will help a potential user understand what the software is used for.

@kthyng I would expect the package documentation to focus on "Howto" with some examples and implementation details. But the underlying _algorithm should be explained in a paper._ I think that the current "Statement of Need" section covers the question of what the software is used for.  The "Algorithm" section is important in my view, matching the policy of "high level".  But the examples are more suited to package documentation.   If you agree with this approach then I'll remove the examples from the paper (leaving them only in the package docs).
