### 2024-07-09 22e50c6
Adds legend to scatter plot for coloring by Feature or Month
