### 2024-07-09 2424179
Larger points in scatterplot
