### 2024-07-09 e98f4cb
Removes examples from paper.md
