+++ inst/paper/paper.md
@@ -88,83 +88,6 @@ and in this case, soil moisture is derived as:
 
 $$ W = \frac{STR - (\alpha_{dry} + \beta1_{dry} \cdot VI + \beta2_{dry} \cdot VI^2)}{(\alpha_{wet} + \beta1_{wet} \cdot VI + \beta2_{wet} \cdot VI^2) - (\alpha_{dry} + \beta1_{dry} \cdot VI + \beta2_{dry} \cdot VI^2)} $$
 
-
-# Examples
-
-### Setup
-
-Define directories and load required packages
-
-``` r
-Output_dir <- tempdir()
-# Edit here...
-work_dir <- system.file("paper", package = "rOPTRAM") 
-GIS_dir <- file.path(work_dir, "GIS")
-aoi_file <- file.path(GIS_dir, "lachish.gpkg")
-
-# load libraries
-remotes::install_github("ropensci/rOPTRAM")
-pkgs <- c("knitr", "ggplot2", "rOPTRAM", "jsonlite")
-invisible(lapply(pkgs, require, character.only = TRUE))
-# model parameters
-veg_index <- "NDVI"
-from_date <- "2022-09-01"
-to_date <- "2023-04-01"
-max_cloud <- 5
-```
-
-### Run model, linear trapezoid fitting
-
-```r
-BOA_list <- optram_acquire_s2(aoi_file = aoi_file,
-                          from_date = from_date,
-                          to_date = to_date, 
-                          max_cloud = max_cloud,
-                          veg_index = veg_index, 
-                          output_dir = Output_dir,
-                          remote = "scihub")
-VI_dir <- file.path(Output_dir, veg_index)
-VI_list <- list.files(VI_dir, full.names = TRUE)
-STR_dir <- file.path(Output_dir, "STR")
-STR_list <- list.files(STR_dir, full.names = TRUE)
-VI_STR_df <- optram_ndvi_str(STR_list, VI_list,
-                             output_dir = Output_dir,
-                             rm.low.vi = TRUE,
-                             rm.hi.str = TRUE)
-
-rmse <- optram_wetdry_coefficients(
-  VI_STR_df,
-  aoi_file = aoi_file,
-  output_dir = GIS_dir,
-  vi_step = 0.005,
-  trapezoid_method = "linear",
-  edge_points = TRUE)
-```
-
-||RMSE wet edge| RMSE dry edge
-|:----------|----------:|----------:|
-|Linear fit: |0.2460|0.0955|
-
-![Linear fitted trapezoid plot.\label{fig:trapezoid_linear}](GIS/trapezoid_lachish_linear.png)
-
-### Second run: polynomial fitted curves
-
-```r
-rmse <- optram_wetdry_coefficients(
-  VI_STR_df,
-  aoi_file = aoi_file,
-  output_dir = GIS_dir,
-  vi_step = 0.005,
-  trapezoid_method = "polynomial",
-  edge_points = TRUE)
-```
-
-||RMSE wet edge|RMSE dry edge|
-|:--------------|---------:|---------:|
-|Polynomial fit:| 0.1490| 0.0931|
-
-![Polynomial fitted trapezoid plot.\label{fig:trapezoid_poly}](GIS/trapezoid_lachish_polynomial.png)
-
 # Acknowledgements
 
 The authors wish to thank the two reviewers, Lauren O'Brien and Harry Eslick, whose comments substantially improved the code in this package for submission to ROpenSci.
