+++ README.Rmd
@@ -1,6 +1,6 @@
 ---
 output: github_document
-bibliography: inst/references.bib
+bibliography: vignettes/references.bib
 ---
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->

[…]
