### 2024-07-10 #6910 @kthyng
Looks good!

Since this has already been reviewed but `rOpenSci`, I have a specific set of steps to go through, which I will start next week probably. Thanks for your patience.
