### 2024-07-26 #6910 @kthyng
@micha-silver Can you link to the rOpenSci review?
