### 2024-07-27 #6910 @micha-silver
@kthyng  Here's the start of the discussion:  [review](https://github.com/ropensci/software-review/issues/612#issue-1940704599)
