### 2024-07-27 #6910 @micha-silver
@kthyng  One reviewer's [comments](https://github.com/ropensci/software-review/issues/612#issuecomment-2027336215)
Second reviewer's [comments](https://github.com/ropensci/software-review/issues/612#issuecomment-2053591926)
