### 2024-08-05 #6910 @kthyng
Here is the `rOpenSci` review: https://github.com/ropensci/software-review/issues/612
