### 2024-08-07 61f8d49
Adds two download and filter options: only-vi-str and tileid
