### 2024-08-07 dbccda2
Fixwes return value of acquire_scihub()
