### 2024-08-07 c67d824
Allows to reset tileid to NA
