+++ man/optram_options.Rd
@@ -49,6 +49,8 @@ feature_col      \tab "ID" \tab string, name of \emph{numeric} column \cr
 edge_points      \tab TRUE  \tab FALSE, whether to add \cr
 \tab  \tab the trapezoid edge points to the plot \cr
 only_vi_str  \tab  FALSE \tab TRUE (avoids downloading all Sentinel bands) \cr
+tileid  \tab  NA \tab a string of 5 characters, \cr
+\tab \tab  download only the requested tileid  \cr
 }
 }
 \examples{
