### 2024-08-07 #7086 @editorialbot
[REVIEW]: `rOPTRAM`: Implementation of the OPTRAM Algorithm in `R`
**Submitting author:** @micha-silver (<a href="http://orcid.org/0000-0002-1128-1325">Micha Silver</a>)
**Repository:** https://github.com/ropensci/rOPTRAM
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
