### 2024-08-07 #7086 @kthyng
Here is the `rOpenSci` review: https://github.com/ropensci/software-review/issues/612
