### 2024-08-07 #7086 @kthyng
@micha-silver So just to be clear we use the review your software already went through with rOpenSci. My next step is to get an archive for your software from something like Zenodo. Looks like you don't have one yet with your rOpenSci review. Could you do that step now, archive the reviewed version of your package with Zenodo or another archiving system and report the DOI here?
