### 2024-08-07 ae305bd
Adds badge with doi on Zenodo
