+++ README.Rmd
@@ -20,6 +20,7 @@ knitr::opts_chunk$set(
 <!-- badges: start -->
 [![Project Status: ACTIVE](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![R-CMD-check](https://github.com/ropensci/rOPTRAM/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ropensci/rOPTRAM/actions/workflows/R-CMD-check.yaml)
+[![DOI](https://zenodo.org/badge/811356946.svg)](https://zenodo.org/doi/10.5281/zenodo.13257766)
 <!-- badges: end -->
 
 `rOPTRAM` implements The OPtical TRapezoid Model (OPTRAM) 
