### 2024-08-07 #7086 @micha-silver
Hi @kthyng
I've added a release on [Zenodo: ](https://zenodo.org/account/settings/github/repository/ropensci/rOPTRAM)
The DOI is [here](https://doi.org/10.5281/zenodo.13257767)

Thanks,
