### 2024-08-08 17b1530
Adds DOI to CITATION
