+++ README.md
@@ -8,6 +8,7 @@
 [![Project Status:
 ACTIVE](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![R-CMD-check](https://github.com/ropensci/rOPTRAM/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ropensci/rOPTRAM/actions/workflows/R-CMD-check.yaml)
+[![DOI](https://zenodo.org/badge/811356946.svg)](https://zenodo.org/doi/10.5281/zenodo.13257766)
 <!-- badges: end -->
 
 `rOPTRAM` implements The OPtical TRapezoid Model (OPTRAM) to derive soil
@@ -161,19 +162,18 @@ rOPTRAM::optram_options()
 #> [1] "feature_col = ID"
 #> [1] "max_cloud = 12"
 #> [1] "max_tbl_size = 1e+06"
+#> [1] "only_vi_str = FALSE"
 #> [1] "period = full"
 #> [1] "plot_colors = no"
 #> [1] "remote = scihub"
 #> [1] "rm.hi.str = FALSE"
 #> [1] "rm.low.vi = FALSE"
 #> [1] "SWIR_band = 11"
+#> [1] "tileid = NA"
 #> [1] "trapezoid_method = linear"
 #> [1] "veg_index = NDVI"
 #> [1] "vi_step = 0.005"
 #> NULL
-```
-
-``` r
 # Change one options
 rOPTRAM::optram_options("veg_index", "SAVI", show_opts = FALSE)
 #> 
