### 2024-08-09 #7086 @kthyng
Ok great! A few things:
* I see the version on Zenodo is different than the version here. Which was reviewed, and therefore the version we should have associated with this JOSS/`rOpenSci` publication?
* Can you update the title and author list on Zenodo to exactly match the JOSS paper?
* Please read through your paper again to make sure everything is perfect since it will be published shortly.
