### 2024-08-10 #7086 @micha-silver
> * I see the version on Zenodo is different than the version here. Which was reviewed, and therefore the version we should have associated with this JOSS/`rOpenSci` publication?

The current version of the package is 0.3.0, the same as the release on Zenodo. Maybe I need some guidance on the nuances of "version", "tag" , and "release".
```
> packageVersion("rOPTRAM")
[1] ‘0.3.0.0’
```
> * Can you update the title and author list on Zenodo to exactly match the JOSS paper?

I see on Zenodo that I can add an "Additional title". Is that sufficient?

Regarding authors, we have some contributors to the package who were not involved in the paper at all. So they are not included. And one author who did not contribute to the package, but is a coauthor of the paper.
??

> * Please read through your paper again to make sure everything is perfect since it will be published shortly.

OK
