### 2024-08-12 #7086 @kthyng
> I see on Zenodo that I can add an "Additional title". Is that sufficient?

> Regarding authors, we have some contributors to the package who were not involved in the paper at all. So they are not included. And one author who did not contribute to the package, but is a coauthor of the paper.

It is not a requirement to change the title and author list to match the Zenodo archive — the idea is to make one cohesive set of archives to represent this review of your software between your `rOpenSci` review, your JOSS DOI, and your Zenodo DOI, thus keeping the same title and author list across them.

The additional title is a different field; if you want to change the metadata to match for this purpose you can directly alter the title and author list.

Yes, it is common for there to be minor contributors to the code who do not rise to the authorship level of the package, but if they should be included then perhaps they should have been included as authors? Just to clarify, the paper itself is a tiny piece of this software publishing and is meant to represent the software package but is not the item being published here — it is the software. Not all authors have to contribute code but they would have contributed to the package in other ways typically like ideas for the software over time.
