### 2024-08-13 #7086 @micha-silver
Hi @kthyng
Thanks for the explanations.

I have changed the Zenodo title, following your suggestion, to match with the title in the package. And, after checking with the other package contributors, I added them as authors to the Zenodo release. So we should be good now.

One clarification: after the review of the package was completed, I added the **reviewers' names** as package contributors (with "role" = "rev"). But they do not appear as authors of the paper in Zenodo. Is that the norm?
