### 2024-08-14 563dfc2
Adds option 'scm_mask' to filter out clouds, cloud shadows, water
