+++ R/zzz.R
@@ -24,7 +24,8 @@
     optram.feature_col = "ID",  # Which column contains numeric ID's of features
     optram.edge_points = TRUE,   # whether to add edge points to the plot
     optram.only_vi_str = FALSE, # whether to avoid downloading all bands
-    optram.tileid = NA # allow to filter download by Sentinel tile ID
+    optram.tileid = NA, # allow to filter download by Sentinel tile ID
+    optram.scm_mask = TRUE # mask out clouds, cloud shadow, snow, water.
   )
 
   op <- options()
