### 2024-08-15 35252ff
Minor edits to JOSS paper
