+++ inst/paper/paper.md
@@ -29,7 +29,7 @@ Soil moisture at the Earth’s surface is essential in many hydrological, pedolo
 
 # Statement of Need
 
-Soil moisture (SM) at the Earth’s surface is essential in many hydrological, pedological, and biological processes. Pasture is the most common land use, where grazing livestock spreads over about 77% of the global agricultural land, and support billions of inhabitants [@ramankutty_farming_2008; @maestre_grazing_2022]. Therefore SM is a critical resource in these regions for the growth of palatable plants, pasture quality, and rangeland health [@mbow_land_2021; @kusserow_desertification_2017]. A time series of SM measurements can serve to determine the quality and potential for grazing. By following seasonal variations in SM, forecasts for preferred grazing lands can be prepared. *In-situ* volumetric soil content is typically measured using Time Domain Reflectometer (TDR) sensors with a high temporal resolution [@kirkham_time_2014]. TDR sensors have been recognized to be very accurate, and can function for long periods. However these data are point measurements, and cannot cover the extensive areas needed for determining regional scale grazing potential. Cosmic Ray Neutron Scanner (CRNS) technology can also acquire SM measurements at high temporal resolution, with a larger spatial scale than TDR, covering up to 30 hectares. These latter instruments are, however, quite expensive, and less accurate than TDR spot measurements [@davies_optimal_2022; @schron_improving_2017]. To overcome the locality issue, Earth Observations from satellites are the most efficient and reliable means for assessing the Earth’s resources, including SM due to its large-scale coverage and repeatability for the last decades. 
+Soil moisture (SM) at the Earth’s surface is essential in many hydrological, pedological, and biological processes. Pasture is the most common land use, where grazing livestock spreads over about 77% of the global agricultural land, and supports billions of inhabitants [@ramankutty_farming_2008; @maestre_grazing_2022]. Therefore SM is a critical resource in these regions for the growth of palatable plants, pasture quality, and rangeland health [@mbow_land_2021; @kusserow_desertification_2017]. A time series of SM measurements can serve to determine the quality and potential for grazing. By following seasonal variations in SM, forecasts for preferred grazing lands can be prepared. *In-situ* volumetric soil content is typically measured using Time Domain Reflectometer (TDR) sensors with a high temporal resolution [@kirkham_time_2014]. TDR sensors have been recognized to be very accurate, and can function for long periods. However these data are point measurements, and cannot cover the extensive areas needed for determining regional scale grazing potential. Cosmic Ray Neutron Scanner (CRNS) technology can also acquire SM measurements at high temporal resolution, with a larger spatial scale than TDR, covering up to 30 hectares. These latter instruments are, however, quite expensive, and less accurate than TDR spot measurements [@davies_optimal_2022; @schron_improving_2017]. To overcome the locality issue, Earth Observations from satellites are the most efficient and reliable means for assessing the Earth’s resources, including SM due to its large-scale coverage and repeatability for the last decades. 
 
 Less than a decade ago, @sadeghi_optical_2017 showed that shortwave infrared (SWIR) Transformed Reflectance (STR) is applicative to assess SM. They developed a novel physical-based model, OPtical TRApezoid Model (OPTRAM) using designated spectral bands of remote sensing imagery. OPTRAM was recently validated [@longo-minnolo_stand-alone_2022] to address the need to estimate SM over vast areas at regional scales. The model can now be applied to various earth observation systems, such as Sentinel-2 or Landsat, with visible, near-infrared, and SWIR bands. [@ambrosone_retrieving_2020; @dubinin_using_2020].
 
@@ -40,7 +40,7 @@ The overarching goal of the current project was to program the original OPTRAM a
 
[…]
