+++ R/acquire_utilities.R
@@ -151,7 +151,15 @@ acquire_scihub <- function(
         time_range <- as.Date(img_list$acquisitionDate[d])
         sourceId <- img_list$sourceId[d]
         tileid <- unlist(strsplit(sourceId, split="_"))[6]
-        result_rast <- CDSE::GetImage(aoi = aoi,
+        raster_file <- file.path(s_dir,
+                                 paste0(basename(s_dir), "_",
+                                        time_range, "_", tileid,
+                                        ".tif"))
+        # Run the GetImage() func only if:
+        # the file has not been downloaded already
+        # or overwrite was set to TRUE
+        if (!file.exists(raster_file) | getOption(optram.overwrite) ) {
+          result_rast <- CDSE::GetImage(aoi = aoi,
                                      time_range = time_range,
                                      script = scrpt,
                                      collection = "sentinel-2-l2a",
@@ -159,12 +167,8 @@ acquire_scihub <- function(
                                      mask = TRUE,
                                      resolution = c(10,10),
                                      token = tok)
-
-        raster_file <- file.path(s_dir,
-                               paste0(basename(s_dir), "_",
-                                      time_range, "_", tileid,
-                                      ".tif"))
-        terra::writeRaster(result_rast, raster_file, overwrite = TRUE)
+          terra::writeRaster(result_rast, raster_file, overwrite = TRUE)
+        }
         return(raster_file)
     })
     return(result_list)
