+++ R/zzz.R
@@ -22,7 +22,8 @@
     optram.rm.hi.str    = FALSE,
     optram.plot_colors = "no",  # "density", "contours", "feature, months"
     optram.feature_col = "ID",  # Which column contains numeric ID's of features
-    optram.edge_points = TRUE   # whether to add edge points to the plot
+    optram.edge_points = TRUE,   # whether to add edge points to the plot
+    optram.overwrite = FALSE    # whether to overwrite previously downloaded S2 files
   )
 
   op <- options()
