### 2024-08-15 2609110
Adds 'overwrite' option to avoid re-downloading previously acquired images
