+++ R/utilities.R
@@ -51,6 +51,7 @@
 #'  \tab \tab  download only the requested tileid  \cr
 #'  scm_mask \tab TRUE \tab FALSE, whether to mask out: \cr
 #'    \tab \tab cloud shadows, clouds, water, snow, using the Copernicus SCM product. \cr
+#'  overwrite \tab  FALSE \tab Set to TRUE to re-download previously acquired images
 #'}
 optram_options <- function(opt_name = NULL, opt_value=NULL,
                            show_opts = TRUE) {
@@ -84,7 +85,7 @@ optram_options <- function(opt_name = NULL, opt_value=NULL,
   opt_names <- c("veg_index","remote", "period", "max_cloud", "vi_step",
                  "trapezoid_method", "SWIR_band", "max_tbl_size",
                  "rm.low.vi", "rm.hi.str", "plot_colors", "feature_col",
-                 "edge_points", "only_vi_str", "tileid", "scm_mask")
+                 "edge_points", "only_vi_str", "tileid", "scm_mask", "overwrite")
   if (opt_name %in% opt_names) {
     # Setup conditions for each option name
     cond_func <- switch(opt_name,
@@ -138,6 +139,9 @@ optram_options <- function(opt_name = NULL, opt_value=NULL,
                          (is.character(opt_value) & nchar(opt_value) == 5))
               },
               "scm_mask" = function(opt_value) {
+                return(is.logical(opt_value))
+              },
+              "overwrite" = function(opt_value) {
                 return(is.logical(opt_value))
               }
       )
