### 2024-08-15 abc4720
Adds 'overwrite' option to avoid redownloading previously acquired imagery
