+++ R/utilities.R
@@ -28,30 +28,30 @@
 #'  opt_name  \tab default  \tab other possible values \cr
 #'  ---------------- \tab---------------\tab ------------------- \cr
 #'  veg_index  \tab "NDVI"  \tab "SAVI", "MSAVI" \cr
-#'  remote  \tab "scihub"   \tab "openeo" \cr
-#'  period  \tab "full"   \tab "seasonal" \cr
-#'  max_cloud \tab 12 \tab between 0 and 100 \cr
-#'  vi_step  \tab 0.005  \tab usually between 0.01 and 0.001 \cr
+#'  remote     \tab "scihub"   \tab "openeo" \cr
+#'  period     \tab "full"   \tab "seasonal" \cr
+#'  max_cloud  \tab 12 \tab between 0 and 100 \cr
+#'  vi_step    \tab 0.005  \tab usually between 0.01 and 0.001 \cr
 #'  trapezoid_method \tab "linear"  \tab "polynomial" or "exponential" \cr
-#'  SWIR_band        \tab 11    \tab 12 \cr
-#'  max_tbl_size     \tab 1e+6  \tab depends on computer resources \cr
-#'  rm.low.vi        \tab FALSE \tab TRUE \cr
-#'  rm.hi.str        \tab FALSE \tab TRUE \cr
-#'  plot_colors      \tab "no"  \tab "no" = uniform green color for all points \cr
+#'  SWIR_band  \tab 11    \tab 12 \cr
+#'  max_tbl_size \tab 1e+6  \tab depends on computer resources \cr
+#'  rm.low.vi   \tab FALSE \tab TRUE \cr
+#'  rm.hi.str   \tab FALSE \tab TRUE \cr
+#'  plot_colors \tab "no"  \tab "no" = uniform green color for all points \cr
 #'   \tab  \tab "features" = points colored by aoi features \cr
 #'   \tab  \tab "density" = points colored by point density \cr
 #'   \tab  \tab "contours" = plots density contour lines \cr
 #'   \tab  \tab "months" = points colored by month of image date \cr
-#'  feature_col      \tab "ID" \tab string, name of *numeric* column \cr
+#'  feature_col \tab "ID" \tab string, name of *numeric* column \cr
 #'   \tab  \tab that contains feature ID's for coloring plot \cr
-#'  edge_points      \tab TRUE  \tab FALSE, whether to add \cr
+#'  edge_points \tab TRUE  \tab FALSE, whether to add \cr
 #'   \tab  \tab the trapezoid edge points to the plot \cr
-#'  only_vi_str  \tab  FALSE \tab TRUE (avoids downloading all Sentinel bands) \cr
-#'  tileid  \tab  NA \tab a string of 5 characters, \cr
-#'  \tab \tab  download only the requested tileid  \cr
-#'  scm_mask \tab TRUE \tab FALSE, whether to mask out: \cr
+#'  only_vi_str \tab  FALSE \tab TRUE (avoids downloading all Sentinel bands) \cr
+#'  tileid      \tab  NA \tab a string of 5 characters, \cr
+#'   \tab \tab  download only the requested tileid  \cr
+#'  scm_mask    \tab TRUE \tab FALSE, whether to mask out: \cr
 #'    \tab \tab cloud shadows, clouds, water, snow, using the Copernicus SCM product. \cr
-#'  overwrite \tab  FALSE \tab Set to TRUE to re-download previously acquired images
+#'  overwrite   \tab  FALSE \tab Set to TRUE to re-download previously acquired images \cr
 #'}
 optram_options <- function(opt_name = NULL, opt_value=NULL,
                            show_opts = TRUE) {
