### 2024-08-15 3561eda
Minor edit to acquire_utilities()
