### 2024-08-15 #7086 @kthyng
> One clarification: after the review of the package was completed, I added the reviewers' names as package contributors (with "role" = "rev"). But they do not appear as authors of the paper in Zenodo. Is that the norm?

If you're asking where they appear on the Zenodo page based on this role assignment, I am not sure, sorry!  But it looks good from the JOSS perspective now. Ready?
