### 2024-08-16 367eee8
Reformat polynomial equation for soil moisture
