### 2024-08-16 748ea3f
Modifies lm() function call for polynomial fit
