### 2024-08-16 #7086 @micha-silver
@kthyng
We made a few last minute grammatical edits. And I also added as an author the package contributor who was missing from the Zenodo release.
So, yes, please go ahead and finalize.
