### 2024-08-18 de0fde0
corrects SAVI_masked.js
