### 2024-08-19 #7086 @kthyng
@micha-silver Did you intend to have your author list different on your paper and Zenodo?
