### 2024-08-19 #7086 @micha-silver
@kthyng No, I liked your suggestion to get the package contributors and paper authors in line. Is there still some discrepancy?
