+++ inst/paper/paper.md
@@ -15,6 +15,11 @@ authors:
     affiliation: 1
   - name: Ron Beiden
     affiliation: 1
+  - name: Zhe Dong
+    orcid: 0009-0007-5298-8355
+    affiliation: 1
+  - name: Natalya Panov
+    affiliation: 1
   - name: Arnon Karnieli
     orcid: 0000-0001-8065-9793
     affiliation: 1
