+++ man/rOPTRAM-package.Rd
@@ -24,6 +24,7 @@ Other contributors:
   \item Arnon Karnieli \email{karnieli@bgu.ac.il} (\href{https://orcid.org/0000-0001-8065-9793}{ORCID}) [contributor, funder]
   \item Ron Beiden \email{beiden@post.bgu.ac.il} [contributor]
   \item Natalya Panov \email{npanov@bgu.ac.il} [contributor]
+  \item Zhe Dong \email{dongzhe053@gmail.com} (\href{https://orcid.org/0009-0007-5298-8355}{ORCID}) [contributor]
   \item Lauren O'Brien \email{obrlsoilau@gmail.com} [reviewer]
   \item Harry Eslick \email{harry@wineandtruffle.com.au} [reviewer]
 }
