### 2024-08-20 #7086 @micha-silver
OK, now I've got 5 contributors, also added as paper authors. (The only exception now is Maelle Salomon from rOpenSci - she helped with a minor code error, so her name was added to the github repository, but **no involvement** with the paper. And as I mentioned, the package reviewers were added to the package, but **not** as paper authors.)
Thanks for all your assistance.
