### 2024-08-22 #7086 @kthyng
Congratulations on your new publication @micha-silver! Many thanks to reviewers @harryeslick and @obrl-soil for your time, hard work, and expertise!! JOSS and `rOpenSci` wouldn't be able to function nor succeed without your efforts.
